refactor scrapers
